v1.0.4-dev1

+++ src/fs/bucket.py
@@ -14,7 +14,7 @@ from urllib.parse import urlparse
 
 class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
     def make(lib_name, name, json_bucket, aeskey):
-        return Bucket(
+        b = Bucket(
             lib_name = lib_name,
             path=json_bucket["path"], 
             max_capacity=json_bucket["max_capacity"],
@@ -24,6 +24,10 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
             aeskey=json_bucket["aeskey"] if "aeskey" in json_bucket else aeskey,
             login=json_bucket["login"] if "login" in json_bucket else '',
             pwd=json_bucket["pwd"] if "pwd" in json_bucket else '')
+        
+        if 'free_capacity' in json_bucket:
+            free_capacity = json_bucket['free_capacity']
+        return b
             
     def __init__(self, lib_name, max_capacity=0, path="", speed=1.0, name="", 
     crypt=False, aeskey="", login='', pwd=''):
@@ -43,17 +47,27 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
         self.login  = login
         self.pwd    = pwd
         
+        
+        dirs = ["data", "locks", "metadata", "transactions"]
+        if not self.scheme: 
+            for d in dirs:
+                tmp = os.path.join(path, d)
+                if not os.path.isdir(tmp):
+                    os.mkdir(tmp)
+        elif self.scheme[:3]=='ftp':
+            with self.connect_ftp() as ftp:
+                for d in dirs:
+                    ftp.mkd(self.urlObj.path+'/'+d if self.urlObj.path else d)
+       
+    def reset_storage(self): #ftp ready
         dirs = ["data", "locks", "metadata", "transactions"] 
         for d in dirs:
-            tmp = os.path.join(path, d)
-            if not os.path.isdir(tmp):
-                os.mkdir(tmp)
-       
-    def reset_storage(self):
-        dirs = ["data", "locks", "metadata", "transactions"] 
-        for d in dirs:
-            shutil.rmtree(os.path.join(self.path, d))
-       
+            if not self.scheme:
+                shutil.rmtree(os.path.join(self.path, d))
+            elif self.scheme[:3]=='ftp':
+                with self.connect_ftp() as ftp:
+                    ftp.rmd(self.urlObj+'/'+d)
+                    
     def access(self, key):
         return self
        
@@ -63,7 +77,7 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
     def buckets(self):
         return [self]
         
-    def write(self, md5, size, fp):
+    def write(self, md5, size, fp):#ftp ready
         with self.lock:
             self.free_capacity -= size
             if self.crypt :
@@ -93,8 +107,8 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
             else:
                 fp2 = fp
                 
-                ftp = self.connect_ftp()
-                ftp.storbinary('STOR '+make_path(self.urlObj.path, self.lib_name, md5), fp2)
+            ftp = self.connect_ftp()
+            ftp.storbinary('STOR '+make_path(self.urlObj.path, self.lib_name, md5), fp2)
             
             if self.crypt :
                 fp2.close()
@@ -102,12 +116,12 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
     
     def connect_ftp(self):
         ftp = FTP() if self.scheme[-1]=='s' else FTP_TLS()
-        ftp.login(self.login+"@"+self.urlObj.netloc, self.pwd)
+        ftp.connect(self.urlObj.netloc)
+        ftp.login(self.login, self.pwd)
         return ftp
     
-    def remove(self, md5, size):
+    def remove(self, md5, size):#ftp ready
         with self.lock:
-            print( self.free_capacity, size, self.max_capacity)
             self.free_capacity += size 
             assert(self.free_capacity <= self.max_capacity)
             
@@ -117,7 +131,7 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
             ftp = self.connect_ftp()
             ftp.delete( make_path(self.urlObj.path, self.lib_name, md5) )
             
-    def get_file(self, md5):
+    def get_file(self, md5):#ftp ready
         if not self.scheme:
             if not self.crypt:
                 return open( make_path(self.path, self.lib_name, md5), "rb")
@@ -151,6 +165,7 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
         state["path"] = self.path
         state["crypt"] = self.crypt
         state["aeskey"] = self.aeskey
+        state["urlObj"] = self.urlObj
         state["scheme"] = self.scheme
         state["login"] = self.login
         state["pwd"] = self.pwd
@@ -163,25 +178,39 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
         self.path = state['path']
         self.crypt = state['crypt']
         self.aeskey = state['aeskey']
+        self.urlObj = state['urlObj']
         self.scheme = state['scheme']
         self.login = state['login']
         self.pwd = state['pwd']
         self.lock=Lock()
 
-    def try_lock(self, idclient, ttl):
-        onlyfiles = [f for f in os.listdir(os.path.join(self.path, "locks")) if os.path.isfile(os.path.join(self.path, f))]
+    def try_lock(self, idclient, ttl):#ftp ready
+        if not self.scheme:
+            onlyfiles = [f for f in os.listdir(os.path.join(self.path, "locks")) if os.path.isfile(os.path.join(self.path, f))]
+        elif self.scheme[:3]=='ftp':
+            with self.ftp_connect() as ftp:
+                onlyfiles = filter(lambda x,y:y['type'] == 'file', ftp.mlsd(self.urlObj.path+'/'+'locks', ['type']))
+                
         for f in onlyfiles:
             if int(f.split("-")[3]) + ttl > time():
                 return False
                 
         self.previous_lock = "lock_%s_%d_%d" % (self.lib_name, idclient, time())
-        fp = open( os.path.join(self.path, "locks", self.previous_lock), 'w')
-        fp.write('0')
         
+        if not self.scheme:
+            fp = open( os.path.join(self.path, "locks", self.previous_lock), 'w')
+            fp.write('0')
+        elif self.scheme[:3]=='ftp':
+            ftp.storbinary('STOR '+self.urlObj.path + '/' + 'locks' + '/' + self.previous_lock, None)
         return True
         
-    def is_locked(self, idclient, ttl):
-        onlyfiles = [f for f in os.listdir(os.path.join(self.path, "locks")) if os.path.isfile(os.path.join(self.path, f))]
+    def is_locked(self, idclient, ttl):#ftp ready
+        if not self.scheme:
+            onlyfiles = [f for f in os.listdir(os.path.join(self.path, "locks")) if os.path.isfile(os.path.join(self.path, f))]
+        elif self.scheme[:3]=='ftp':
+            with self.ftp_connect() as ftp:
+                onlyfiles = filter(lambda x,y:y['type'] == 'file', ftp.mlsd(self.urlObj.path+'/'+'locks', ['type']))
+                
         onlyfiles.filter(lambda f:f.split('-')[1]==self.lib_name)
         onlyfiles = sorted(onlyfiles, keys=lambda x: int(x.split('-')[3]), reverse=True)
         
@@ -194,22 +223,40 @@ class Bucket(Container): #décrit un dossier ex: photo, gdrive, dropbox
             True
         
     
-    def unlock(self):
-        if os.path.exists(os.path.join(self.path, "locks", self.previous_lock)):
-            os.remove(os.path.join(self.path, "locks", self.previous_lock))
-        
+    def unlock(self):#ftp ready
+        if not self.scheme:
+            if os.path.exists(os.path.join(self.path, "locks", self.previous_lock)):
+                os.remove(os.path.join(self.path, "locks", self.previous_lock))
+        elif self.scheme[:3]=='ftp':
+            with self.ftp_connect() as ftp:
+                onlyfiles = filter(lambda x,y:y['type'] == 'file', ftp.mlsd(self.urlObj.path+'/'+'locks', ['type']))
+                if self.previous_lock in map(lambda x,y:x, onlyfiles):
+                    ftp.delete(self.urlObj.path+'/'+'locks'+'/'+self.previous_lock)
+
         self.previous_lock = ''
         
-    def transactions(self):#todo for ftp
+    def transactions(self):#ftp ready
         transactions = set()
-        for path, dirs, files in os.walk(os.path.join(self.path, 'transactions')): #rewirte with scandir for 3.5
-            for filename in files:
-                location = os.path.join(path, filename)
-                
-                id_lib, id_client, _ = filename.split('-')
-                if(id_lib == self.lib_name):
-                    tmp = Transaction(id_client, id_lib)
-                    tmp.load(os.path.join(path, filename))
-                    transactions.add( tmp )
-        
+
+        if not self.scheme:
+            for path, dirs, files in os.walk(os.path.join(self.path, 'transactions')): #rewirte with scandir for 3.5
+                for filename in files:
+                    location = os.path.join(path, filename)
+                    
+                    id_lib, id_client, _ = filename.split('-')
+                    if(id_lib == self.lib_name):
+                        tmp = Transaction(id_client, id_lib)
+                        tmp.load(os.path.join(path, filename))
+                        transactions.add( tmp )
+        elif self.scheme[:3]=='ftp':
+            with self.ftp_connect() as ftp:
+                onlyfiles = filter(lambda x,y:y['type'] == 'file', ftp.mlsd(self.urlObj.path+'/'+'transactions', ['type']))
+                for filename in onlyfiles:
+                    location = os.path.join(path, filename)
+                    
+                    id_lib, id_client, _ = filename.split('-')
+                    if(id_lib == self.lib_name):
+                        tmp = Transaction(id_client, id_lib)
+                        tmp.load(os.path.join(path, filename))
+                        transactions.add( tmp )
         return transactions
